RENAMED FRONTOFFICE AND BACKOFFICE TO FRONTEND AND BACKEND
I renamed the backoffice folder to backend and the frontoffice folder to
frontend. I also edited all the files so that i keeps working with the new
names.
It is important to name correctly the folders. Backoffice and backend are
different things. Correct naming will be important if new people join the team.