ADDED ROUTE FOR COMMUNITY, CHANGED DATA MODEL OF COMMUNITY & CREATED COMMUNITY PROVIDER
I changed the data model of community to add list of currentEvents in the
community and the list of passedEvents in the community.
I also changed the route of postCommuniti to add the two new events fields.
I also created a provider in the frontend to retrieve community depending on
community_id in the database.